PaymentCompleteService  - 결제 완료 기능 추가 및 테스트 검증

+++ payment-service/src/test/java/com/example/backend/application/test/R2DBCPaymentDatabaseHelper.java
@@ -55,7 +55,7 @@ public class R2DBCPaymentDatabaseHelper implements PaymentDatabaseHelper {
                                                         .orderId((String) resultMap.get("order_id"))
                                                         .productId((Long) resultMap.get("product_id"))
                                                         .sellerId((Long) resultMap.get("seller_id"))
-                                                        .amount((Long) resultMap.get("amount"))
+                                                        .amount(((BigDecimal) resultMap.get("amount")).longValue())
                                                         .paymentStatus(PaymentStatus.valueOf((String) resultMap.get("order_status")))
                                                         .isLedgerUpdated(isLedgerUpdated)
                                                         .isWalletUpdated(isWalletUpdated)
